迁移ID类型至INT。单元测试。

+++ src/main/java/org/example/express_backend/dto/CalculatePriceDTO.java
@@ -6,8 +6,8 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class CalculatePriceDTO {
-    private Integer origin; // 出发地编号
-    private Integer destination;    // 目的地编号
+    private Long origin; // 出发地编号
+    private Long destination;    // 目的地编号
     private Double weight;  // 包裹重量，单位kg
     private String size;    // 包裹尺寸，如 20,30,40 单位cm
     private Integer type;   // 包裹类型，0为标快，1为特快
