+++ src/main/java/org/example/express_backend/dto/CreateBatchDTO.java
@@ -6,8 +6,8 @@ import java.io.Serializable;
 
 @Data
 public class CreateBatchDTO implements Serializable {
-    private Integer origin; // 出发地转运中心编号
-    private Integer destination;    // 目的地转运中心编号
-    private Integer responsible;    // 负责人id
-    private Integer vehicleId;  // 载具id
+    private Long origin; // 出发地转运中心编号
+    private Long destination;    // 目的地转运中心编号
+    private Long responsible;    // 负责人id
+    private Long vehicleId;  // 载具id
 }
