+++ src/main/java/org/example/express_backend/dto/VehicleDto.java
@@ -7,6 +7,6 @@ import org.example.express_backend.entity.Point;
 @Data
 @ToString
 public class VehicleDto {
-    private Integer id; // 载具id
+    private Long id; // 载具id
     private Point coordinate;   // 坐标
 }
